Support color remapping for rubiks cube pictures
The idea is to produce a dither with the N best colors, then remap those
colors to some from another palette. The hope is that a more interesting
dithering pattern is produced at the cost of shifting the original colors
in the image by some amount. It works, but could probably be improved.